Moved the app's navigation menu to its own component.
