+++ src/app/app.component.spec.ts
@@ -44,15 +44,11 @@ import {
 } from "@angular/platform-browser-dynamic/testing";
 import {} from "jasmine";
 import { APP_BASE_HREF, CommonModule } from "@angular/common";
-import { ReactiveFormsModule } from "@angular/forms";
-import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { BehaviorSubject, of, throwError } from "rxjs";
 import { provideZoneChangeDetection, signal } from "@angular/core";
 import { MockComponent, MockProvider } from "ng-mocks";
-import { setViewport } from "@web/test-runner-commands";
 
 import { AppComponent } from "./app.component";
-import { NotificationsTab } from "./components/notifications/notifications.component";
 import { AuthService } from "@app/services/auth.service";
 import { SWManager } from "@app/services/sWManager.service";
 import { NotificationService } from "./services/notifications.service";
@@ -60,10 +56,11 @@ import { mockAuthedUser } from "@tests/mockData";
 import { AppAlert } from "./components/appAlert/appAlert.component";
 import { AlertsService } from "@app/services/alerts.service";
 import { ItemsService } from "./services/items.service";
+import { AppNavMenu } from "./components/layout/navigationMenu/navigationMenu.component";
 
 describe("AppComponent", () => {
   beforeEach(() => {
-    const MockNotificationsTab = MockComponent(NotificationsTab);
+    const MockNavBar = MockComponent(AppNavMenu);
     const MockAppAlert = MockComponent(AppAlert);
     const MockAuthService = MockProvider(AuthService, {
       authenticated: signal(true),
@@ -90,16 +87,7 @@ describe("AppComponent", () => {
     TestBed.initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting());
 
     TestBed.configureTestingModule({
-      imports: [
-        CommonModule,
-        RouterOutlet,
-        RouterLink,
-        ReactiveFormsModule,
-        FontAwesomeModule,
-        MockNotificationsTab,
-        MockAppAlert,
-        AppComponent,
-      ],
+      imports: [CommonModule, RouterOutlet, RouterLink, MockAppAlert, MockNavBar, AppComponent],
       providers: [
         { provide: APP_BASE_HREF, useValue: "/" },
         provideZoneChangeDetection({ eventCoalescing: true }),
@@ -259,381 +247,6 @@ describe("AppComponent", () => {
     expect(navigateSpy).toHaveBeenCalledWith(["/test"]);
     expect(paramMapSpy).toHaveBeenCalledWith("redirect");
   }));
-
-  // Check that there are valid navigation links
-  it("should contain valid navigation links", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const componentHtml = fixture.debugElement.nativeElement;
-
-    let navMenu = componentHtml.querySelector("#navLinks");
-    expect(navMenu).toBeDefined();
-    expect(navMenu!.children.length).not.toBe(0);
-
-    // check each navingation item to ensure it contains a link
-    let navMenuItems = navMenu!.children;
-    for (var i = 0; i < navMenuItems.length; i++) {
-      expect(navMenuItems.item(i)).toBeDefined();
-      expect(navMenuItems!.item(i)!.children.item(0)!.getAttribute("href")).toBeDefined();
-      expect(navMenuItems!.item(i)!.children.item(0)!.getAttribute("href")).not.toBe("");
-    }
-  });
-
-  // Check that the notifications tab is hidden
-  it("has hidden notifications tab", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.debugElement.nativeElement;
-    fixture.detectChanges();
-
-    expect(component.showNotifications()).toBe(false);
-    expect(componentHtml.querySelector("#mainContent").children.length).toEqual(2);
-  });
-
-  // Check that the notifications tab appears when the button is clicked
-  it("has a notifications tab that appears when its icon is clicked", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const mainContent = componentHtml.querySelector("#mainContent");
-
-    // Check the tab is initially hidden
-    expect(component.showNotifications()).toBe(false);
-    expect(mainContent.querySelector("app-notifications")).toBeNull();
-
-    // Simulate a click on the button
-    componentHtml.querySelector("#notificationsBtn").click();
-    fixture.detectChanges();
-
-    // Check the tab is now visible
-    expect(component.showNotifications()).toBe(true);
-    expect(mainContent.querySelector("app-notifications")).toBeDefined();
-  });
-
-  // Check that the search panel is hidden
-  it("has hidden search", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.debugElement.nativeElement;
-
-    expect(component.showSearch()).toBe(false);
-    expect(componentHtml.querySelector("#siteHeader").children.length).toEqual(2);
-  });
-
-  // Check that the search panel appears when the button is clicked
-  it("has a search which appears when the icon is clicked", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const siteHeader = componentHtml.querySelector("#siteHeader");
-
-    // Check the panel is initially hidden
-    expect(component.showSearch()).toBe(false);
-    expect(siteHeader.querySelector("#search")).toBeNull();
-
-    // Simulate a click on the button
-    componentHtml.querySelector("#searchBtn").click();
-    fixture.detectChanges();
-
-    // Check the panel is now visible
-    expect(component.showSearch()).toBe(true);
-    expect(siteHeader.querySelector("#search")).toBeDefined();
-  });
-
-  // Check that clicking 'search' triggers the ItemsService
-  it("should pass search query to the ItemsService when clicking search", (done: DoneFn) => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const searchSpy = spyOn(component, "searchApp").and.callThrough();
-    const searchServiceSpy = spyOn(component["itemsService"], "sendSearch");
-    spyOn(component["router"], "navigate");
-
-    // open search panel and run search
-    componentHtml.querySelector("#searchBtn").click();
-    fixture.detectChanges();
-    // tick();
-    componentHtml.querySelector("#searchQuery").value = "search";
-    componentHtml.querySelector("#searchQuery").dispatchEvent(new Event("input"));
-    componentHtml.querySelectorAll(".sendData")[0].click();
-
-    // check the spies were triggered
-    expect(searchSpy).toHaveBeenCalled();
-    expect(searchServiceSpy).toHaveBeenCalled();
-    expect(searchServiceSpy).toHaveBeenCalledWith("search");
-    done();
-  });
-
-  // Check that an empty search query isn't allowed
-  it("should prevent empty searches", (done: DoneFn) => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const searchSpy = spyOn(component, "searchApp").and.callThrough();
-    const searchServiceSpy = spyOn(component["itemsService"], "sendSearch");
-    const alertsSpy = spyOn(component["alertsService"], "createAlert");
-
-    // open search panel and run search
-    componentHtml.querySelector("#searchBtn").click();
-    fixture.detectChanges();
-
-    componentHtml.querySelector("#searchQuery").value = "";
-    componentHtml.querySelectorAll(".sendData")[0].click();
-
-    // check one spy was triggered and one wasn't
-    expect(searchSpy).toHaveBeenCalled();
-    expect(searchServiceSpy).not.toHaveBeenCalled();
-    expect(alertsSpy).toHaveBeenCalledWith({
-      message: "Search query is empty! Please write a term to search for.",
-      type: "Error",
-    });
-    done();
-  });
-
-  // Check that the font size panel is hidden
-  it("should have a hidden font size panel", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.debugElement.nativeElement;
-
-    expect(component.showTextPanel()).toBe(false);
-    expect(componentHtml.querySelector("#siteHeader").children.length).toEqual(2);
-  });
-
-  // Check that the font size panel appears when the button is clicked
-  it("has a font size which appears when the icon is clicked", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const siteHeader = componentHtml.querySelector("#siteHeader");
-
-    // Check the panel is initially hidden
-    expect(component.showTextPanel()).toBe(false);
-    expect(siteHeader.querySelector("#textPanel")).toBeNull();
-
-    // Simulate a click on the button
-    componentHtml.querySelector("#textSize").click();
-    fixture.detectChanges();
-
-    // Check the panel is now visible
-    expect(component.showTextPanel()).toBe(true);
-    expect(siteHeader.querySelector("#textPanel")).toBeDefined();
-  });
-
-  // Check that the font size panel is hidden when the button is clicked again
-  it("has a font size which is hidden when the icon is clicked again", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const siteHeader = componentHtml.querySelector("#siteHeader");
-
-    // Check the panel is initially hidden
-    expect(component.showTextPanel()).toBe(false);
-    expect(siteHeader.querySelector("#textPanel")).toBeNull();
-
-    // Simulate a click on the button
-    componentHtml.querySelector("#textSize").click();
-
-    // Check the panel is now visible
-    expect(component.showTextPanel()).toBe(true);
-    expect(siteHeader.querySelector("#textPanel")).toBeDefined();
-
-    // Simulate another click on the button
-    componentHtml.querySelector("#textSize").click();
-
-    // check the panel is hidden again
-    expect(component.showTextPanel()).toBe(false);
-    expect(siteHeader.querySelector("#textPanel")).toBeNull();
-  });
-
-  // Check that the font size panel changes the site's font size
-  it("has a font size that changes according to user choice", (done: DoneFn) => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const fontButton = componentHtml.querySelector("#textSize");
-    const menuSpy = spyOn(component, "checkMenuSize");
-
-    // open the text panel
-    fontButton.click();
-    fixture.detectChanges();
-
-    const fontPanelButtons = componentHtml
-      .querySelector("#textPanel")
-      .querySelectorAll(".appButton");
-
-    // wrap tests in a promise to make sure they run fully and by the order
-    // step 1: regular size
-    new Promise((resolve) => {
-      // change the font size to the smallest
-      fontPanelButtons[0]!.click();
-      fixture.detectChanges();
-
-      // check the font size was changed
-      expect(document.querySelector("html")!.style.fontSize).toBe("75%");
-      expect(menuSpy).toHaveBeenCalled();
-      expect(menuSpy).toHaveBeenCalledTimes(1);
-      resolve(undefined);
-      // step 3: smaller size
-    })
-      .then(() => {
-        // change the font size to the smaller
-        fontPanelButtons[1]!.click();
-        fixture.detectChanges();
-
-        // check the font size was changed
-        expect(document.querySelector("html")!.style.fontSize).toBe("87.5%");
-        expect(menuSpy).toHaveBeenCalled();
-        expect(menuSpy).toHaveBeenCalledTimes(2);
-        // step 4: regular size
-      })
-      .then(() => {
-        // change the font size to the normal
-        fontPanelButtons[2]!.click();
-        fixture.detectChanges();
-
-        // check the font size was changed
-        expect(document.querySelector("html")!.style.fontSize).toBe("100%");
-        expect(menuSpy).toHaveBeenCalled();
-        expect(menuSpy).toHaveBeenCalledTimes(3);
-        // step 5: larger size
-      })
-      .then(() => {
-        // change the font size to the larger
-        fontPanelButtons[3]!.click();
-        fixture.detectChanges();
-
-        // check the font size was changed
-        expect(document.querySelector("html")!.style.fontSize).toBe("150%");
-        expect(menuSpy).toHaveBeenCalled();
-        expect(menuSpy).toHaveBeenCalledTimes(4);
-        // step 6: largest size
-      })
-      .then(() => {
-        // change the font size to the largest
-        fontPanelButtons[4]!.click();
-        fixture.detectChanges();
-
-        // check the font size was changed
-        expect(document.querySelector("html")!.style.fontSize).toBe("200%");
-        expect(menuSpy).toHaveBeenCalled();
-        expect(menuSpy).toHaveBeenCalledTimes(5);
-        done();
-      });
-  });
-
-  // check the menu is shown if the screen is wide enough
-  it("should show the menu if the screen is wide enough", async () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    await setViewport({ width: 700, height: 640 });
-    fixture.detectChanges();
-
-    expect(component.showMenu()).toBeTrue();
-    expect(componentHtml.querySelector("#navLinks")!.classList).not.toContain("hidden");
-    expect(componentHtml.querySelector("#menuBtn")!.classList).toContain("hidden");
-  });
-
-  // TODO: Figure out why this isn't working in CI.
-  // check the menu is hidden if the screen isn't wide enough
-  // it("should hide the menu if the screen isn't wide enough", async () => {
-  //   const fixture = TestBed.createComponent(AppComponent);
-  //   const component = fixture.componentInstance;
-  //   const componentHtml = fixture.nativeElement;
-  //   await setViewport({ width: 600, height: 640 });
-  //   fixture.detectChanges();
-
-  //   expect(component.showMenu()).toBeFalse();
-  //   expect(componentHtml.querySelector("#navLinks")!.classList).toContain("hidden");
-  // });
-
-  // check the menu is hidden when clicked again
-  // it("should show/hide the menu when the menu button is clicked", async () => {
-  //   const fixture = TestBed.createComponent(AppComponent);
-  //   const component = fixture.componentInstance;
-  //   const componentHtml = fixture.nativeElement;
-  //   await setViewport({ width: 600, height: 640 });
-  //   fixture.detectChanges();
-
-  //   // pre-click check
-  //   expect(component.showMenu()).toBeFalse();
-  //   expect(componentHtml.querySelector("#navLinks")!.classList).toContain("hidden");
-  //   expect(componentHtml.querySelector("#menuBtn")!.classList).not.toContain("hidden");
-
-  //   // trigger click
-  //   componentHtml.querySelector("#menuBtn").click();
-  //   fixture.detectChanges();
-
-  //   // post-click check
-  //   expect(component.showMenu()).toBeTrue();
-  //   expect(componentHtml.querySelector("#navLinks")!.classList).not.toContain("hidden");
-  //   expect(componentHtml.querySelector("#menuBtn")!.classList).not.toContain("hidden");
-
-  //   // trigger another click
-  //   componentHtml.querySelector("#menuBtn").click();
-  //   fixture.detectChanges();
-
-  //   // post-click check
-  //   expect(component.showMenu()).toBeFalse();
-  //   expect(componentHtml.querySelector("#navLinks")!.classList).toContain("hidden");
-  //   expect(componentHtml.querySelector("#menuBtn")!.classList).not.toContain("hidden");
-  // });
-
-  // should hide the nav menu if it gets too long
-  it("changeTextSize - should hide nav menu if it gets too long", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const checkSpy = spyOn(component, "checkMenuSize").and.callThrough();
-    fixture.detectChanges();
-
-    const navMenu = componentHtml.querySelector("#navMenu");
-    const navLinks = componentHtml.querySelector("#navLinks");
-    navLinks.style.width = "600px";
-    navMenu.style.maxWidth = "600px";
-    navMenu.style.display = "flex";
-    component.changeTextSize("largest");
-    fixture.detectChanges();
-
-    expect(checkSpy).toHaveBeenCalled();
-    expect(navLinks.classList).toContain("hidden");
-    expect(componentHtml.querySelector("#menuBtn").classList).not.toContain("hidden");
-  });
-
-  // should hide the menu if it gets too long and show it again if it's not too long
-  it("should show the menu again if it's not too long again", () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const component = fixture.componentInstance;
-    const componentHtml = fixture.nativeElement;
-    const checkSpy = spyOn(component, "checkMenuSize").and.callThrough();
-    fixture.detectChanges();
-
-    const navMenu = componentHtml.querySelector("#navMenu");
-    const navLinks = componentHtml.querySelector("#navLinks");
-    navLinks.style.width = "1000px";
-    navMenu.style.maxWidth = "1000px";
-    navMenu.style.display = "flex";
-    component.changeTextSize("largest");
-    fixture.detectChanges();
-
-    // Validate it's hidden before un-hiding it
-    expect(componentHtml.querySelector("#menuBtn").classList).not.toContain("hidden");
-
-    navLinks.style.width = "500px";
-    component.changeTextSize("smaller");
-    fixture.detectChanges();
-
-    expect(checkSpy).toHaveBeenCalled();
-    expect(navLinks.classList).not.toContain("hidden");
-    expect(component.showMenu()).toBeTrue();
-  });
 
   // check the 'share' button is hidden
   it("shouldn't show the share button if it's not supported", () => {
