+++ src/app/components/layout/navigationMenu/navigationMenu.component.less
@@ -1,0 +1,239 @@
+/* Site logo */
+#siteLogo {
+  margin: 0 10px 0 0;
+  padding: 0 10px 0 0;
+  height: 100px;
+  width: 100px;
+  min-width: 60px;
+  /* Responsive styling for screens above 600px width */
+  @media screen and (min-width: 600px) and (max-width: 750px) {
+    height: 75px;
+    width: 75px;
+  }
+  /* Responsive styling for screens between 750px and 800px width */
+  @media screen and (min-width: 750px) and (max-width: 800px) {
+    height: 90px;
+    width: 90px;
+    margin: 0 15px 0 0;
+  }
+  /* Responsive styling for screens above 800px width */
+  @media screen and (min-width: 800px) {
+    padding: 0 15px 0 0;
+  }
+}
+
+/* Navigation Menu */
+#navMenu {
+  width: 100%;
+  max-width: 100%;
+  left: 0;
+  top: 0;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-bottom: thin solid #cfcfcf;
+  /* Responsive styling for screens up to 400px width */
+  @media screen and (max-width: 400px) {
+    padding: 0;
+  }
+  /* Responsive styling for screens between 400px and 650px width */
+  @media screen and (min-width: 400px) and (max-width: 650px) {
+    width: 90%;
+    max-width: 90%;
+    margin: 0 5%;
+  }
+  /* Responsive styling for screens above 650px width */
+  @media screen and (min-width: 650px) {
+    margin: 0;
+    padding: 10px 0 0;
+  }
+}
+
+/* Menu button icon */
+.fa-bars {
+  font-size: 24px;
+  text-align: center;
+}
+
+/* Navigation item icon */
+.navIcon {
+  font-size: 30px;
+  text-align: center;
+  /* Responsive styling for screens above 650px width */
+  @media screen and (min-width: 650px) and (max-width: 700px) {
+    font-size: 26px;
+  }
+  /* Responsive styling for screens between 700px and 800px width */
+  @media screen and (min-width: 700px) and (max-width: 800px) {
+    font-size: 28px;
+  }
+}
+
+/* Navigation Item in the large menu */
+.navLinks.large .navLink,
+.navLinks.large .navButton {
+  padding: 10px 15px;
+  background: var(--boxesBackground);
+}
+
+/* Navigation item image (for those that have it) */
+.navLink img {
+  padding: 5px 5px 6px;
+}
+
+/* Navigation Item (hover) */
+.navLink:hover,
+.navLink.active:hover {
+  color: #a50d0d;
+}
+
+/* Navigation Item (current page) */
+.navLink.active {
+  color: #9d154e;
+}
+
+/* Navigation item pic (when it's an SVG rather than font awesome) */
+.navPic {
+  height: 30px;
+  padding: 7px 5px 4px;
+  text-align: center;
+  /* Responsive styling for screens above 650px width */
+  @media screen and (min-width: 650px) and (max-width: 700px) {
+    height: 26px;
+  }
+  /* Responsive styling for screens between 700px and 800px width */
+  @media screen and (min-width: 700px) and (max-width: 800px) {
+    height: 28px;
+  }
+}
+
+/* Notifications button - badge */
+.newNotification .badge {
+  right: -5px;
+}
+
+/* Notifications button icon */
+#navNotific {
+  /* Responsive styling for screens up to 650px width */
+  @media screen and (max-width: 650px) {
+    padding: 2px;
+  }
+}
+
+.newNotification .badge {
+  background: #cfcfcf;
+}
+
+/* Text size button icon */
+#textSize {
+  margin: 2px;
+  padding: 2px;
+}
+
+/* Search form / font size adjustment */
+#search,
+#textPanel {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  width: 100%;
+  max-width: 100%;
+  margin: 10px 0 5px;
+  left: 0;
+  right: 0;
+  border-radius: 5px;
+  box-shadow: 3px 3px 7px 0 rgba(46, 61, 73, 0.5);
+  position: relative;
+  /* Responsive design for screens above 400px width */
+  @media screen and (min-width: 400px) {
+    width: 90%;
+    max-width: 90%;
+    margin: 10px 5% 5px;
+  }
+  /* Responsive design for screens above 450px width */
+  @media screen and (min-width: 450px) {
+    justify-content: space-evenly;
+    padding: 0 20px;
+    width: calc(90% - 20px);
+    max-width: calc(90% - 20px);
+    margin: 10px calc(5% - 10px) 5px;
+  }
+  /* Responsive design for screens above 525px width */
+  @media screen and (min-width: 525px) {
+    width: calc(84% - 20px);
+    max-width: calc(84% - 20px);
+    margin: 10px calc(8% - 10px) 5px;
+  }
+  /* Responsive design for screens above 575px width */
+  @media screen and (min-width: 575px) {
+    width: calc(80% - 20px);
+    max-width: calc(80% - 20px);
+    margin: 10px calc(10% - 10px) 5px;
+  }
+  /* Responsive design for screens above 625px width */
+  @media screen and (min-width: 625px) {
+    width: calc(76% - 20px);
+    max-width: calc(76% - 20px);
+    margin: 10px calc(12% - 10px) 5px;
+  }
+  /* Responsive design for screens above 650px width */
+  @media screen and (min-width: 650px) {
+    justify-content: center;
+    width: calc(70% - 20px);
+    max-width: calc(70% - 20px);
+    margin: 10px calc(15% - 10px) 5px;
+  }
+  /* Responsive design for screens above 700px width */
+  @media screen and (min-width: 700px) {
+    width: calc(60% - 20px);
+    max-width: calc(60% - 20px);
+    margin: 10px calc(20% - 10px) 5px;
+  }
+}
+
+/* font size adjustment panel */
+#textPanel .appButton {
+  margin: 5px;
+  font-weight: normal;
+}
+
+/* Search label */
+#searchLabel {
+  order: 1;
+  left: 0;
+  margin: 0;
+  align-self: center;
+  padding: 10px 0;
+  /* Responsive design for screens above 400px width */
+  @media screen and (min-width: 400px) {
+    margin: 0 7px;
+  }
+}
+
+/* Search field */
+#searchQuery {
+  order: 2;
+  padding: 0;
+  margin: 2px 3px;
+  height: 20px;
+  align-self: center;
+  /* Responsive design for screens above 400px width */
+  @media screen and (min-width: 400px) {
+    margin: 2px 5px;
+  }
+  /* Responsive design for screens above 550px width */
+  @media screen and (min-width: 550px) {
+    margin: 2px 10px;
+  }
+}
+
+/* Search button */
+#search .sendData {
+  order: 3;
+  margin: 1px;
+  align-self: center;
+  /* Responsive design for screens above 400px width */
+  @media screen and (min-width: 400px) {
+    margin: 1px 2px;
+  }
+}
